Move the validation fuzz targets to the package that owns the code
ParseImageRef, CheckID, CheckContainerName, CheckPortKey and CheckHostPort
moved to dockerclient when the interface was split out, but their fuzz
targets stayed behind in dockerapi. Someone reading dockerclient's tests
would conclude the validators were never fuzzed, and a second
implementation would not inherit the coverage.

The four targets move across and now call the exported names. The
dockerapi-specific ones stay where they are: parseHost, compareVersions,
demux, drainPullStream and newStatusError are that package's own code, and
FuzzValidateAndWriteTar deliberately pairs dockerclient's validator with
dockerapi's tar writer, because the bug it found was a disagreement
between the two.

Each target holds for 20s of fuzzing after the move.

Also records in PLAN.md why go.work cannot be committed yet: a workspace
resolves one version of github.com/docker/go-connections across every
member, the legacy root client needs v0.4.0 and moby/moby/client needs
v0.7.0, and v0.7.0 dropped the symbol the old client calls. The workspace
has to wait until the old implementation is deleted.
